Template & Test Rearrangement (#23)
* create an Organization model

* add basic test for Organization model

* show Organizations in admin

* org dashboard page shows user's Organizations

* user may edit or create Organization

* remove unused test file

* user may not edit Organizations they are associated with

* user may delete an Organization that they are associated with

* user does not edit or create slug for Organizations; unique one is created on first save

* users must be authenticated to create, edit, delete Organizations

* creating an Organization automatically associates creator user with it

* define a base test case mixin

* add a OrgResourceAccess model to keep track of which Organization has access to a user's resource

* show OrgResourceAccess instances in admin

* organization dashboard lists which Organizations have access to which resources for a user

* rename OrgResourceAccess model to ResourceGrant

* rename 'org_access_to_providers' to 'resource_grants_for_user'

* rename 'org_access_to_providers' to 'resource_grants_for_user' in template

* give Organization model 'created', 'updated' fields, which are set automatically

* ResourceGrant references python wrapper for UserSocialAuth model

* display relevant ResourceGrants on org dashboard page

* Style login page according to prototype

TODO:
- [ ] scss pipeline
- [ ] partials for things like `head`

* Provide tooling for building scss

* ResourceGrant.provider_name is now a property

* Resource class defines its name, rather than relying on name to be passed to its methods

* update use of Resource class's filter method to its new name

* give ResourceGrant model 'created', 'updated' fields, which are set automatically

* create a common class for keeping track of when an object is created and updated

* move CreatedUpdatedModel to common app

* squash migrations

* add django-phonenumber-field library to repo for phone number handling

* add 'phone' field to Organization model

* add localflavor library to repo for address validation

* add address fields to Organization model

* move template files into subdirectories, so they are not accidentally found by other apps

* define common tests class and factory in common app
