+++ apps/member/views.py
@@ -3,4 +3,4 @@ from django.views.generic.base import TemplateView
 
 
 class DashboardView(LoginRequiredMixin, TemplateView):
-    template_name = "dashboard.html"
+    template_name = "member/dashboard.html"
