+++ apps/org/tests/factories.py
@@ -1,9 +1,9 @@
 import random
 
-from django.conf import settings
 from django.template.defaultfilters import slugify
 from factory import DjangoModelFactory, Faker, LazyAttribute, SubFactory
 
+from apps.common.tests.factories import UserFactory
 from ..models import Organization, ResourceGrant
 
 
@@ -19,16 +19,6 @@ class OrganizationFactory(DjangoModelFactory):
 
     class Meta:
         model = Organization
-
-
-class UserFactory(DjangoModelFactory):
-    first_name = Faker('first_name')
-    last_name = Faker('last_name')
-    email = Faker('email')
-    username = Faker('user_name')
-
-    class Meta:
-        model = settings.AUTH_USER_MODEL
 
 
 class UserSocialAuthFactory(DjangoModelFactory):
