+++ apps/org/views.py
@@ -10,7 +10,7 @@ from .models import Organization, ResourceGrant
 
 
 class DashboardView(LoginRequiredMixin, TemplateView):
-    template_name = "dashboard.html"
+    template_name = "org/dashboard.html"
 
     def get_context_data(self, **kwargs):
         """Add the user's Organizations to the context."""
@@ -43,7 +43,7 @@ class DashboardView(LoginRequiredMixin, TemplateView):
 class CreateOrganizationView(LoginRequiredMixin, CreateView):
     model = Organization
     fields = ['name', 'users']
-    template_name = 'organization.html'
+    template_name = 'org/organization.html'
     success_url = reverse_lazy('org:dashboard')
 
     def form_valid(self, form):
@@ -58,7 +58,7 @@ class CreateOrganizationView(LoginRequiredMixin, CreateView):
 class UpdateOrganizationView(LoginRequiredMixin, UpdateView):
     model = Organization
     fields = ['name', 'users']
-    template_name = 'organization.html'
+    template_name = 'org/organization.html'
     success_url = reverse_lazy('org:dashboard')
 
     def get_queryset(self):
@@ -70,7 +70,7 @@ class UpdateOrganizationView(LoginRequiredMixin, UpdateView):
 class DeleteOrganizationView(LoginRequiredMixin, DeleteView):
     model = Organization
     success_url = reverse_lazy('organization-list')
-    template_name = 'organization_confirm_delete.html'
+    template_name = 'org/organization_confirm_delete.html'
     success_url = reverse_lazy('org:dashboard')
 
     def get_queryset(self):
